Change the name of the class Added a bunch more tests. More to add
